create parents class GEneticAlgorithm and PopulationRepertoire
